refactor: centralize all NIRSpec plotting code in plots.py
Move all matplotlib usage from stage modules into a single plots.py
module, eliminating scattered module-level matplotlib imports that
caused interactive backend errors on headless Linux.

New functions in plots.py:
- plot_bkg_subtraction: background subtraction diagnostic (from stage1)
- plot_extraction_profiles: 4-panel profile comparison (deduplicated
  from two identical copies in stage3)
- plot_spectrum_qa: 3-row spectrum diagnostic (deduplicated from two
  near-identical copies in stage3)
- plot_stuck_shutter_diagnostics: stuck shutter QA (from stuck_shutters)

Also adds shared style infrastructure (_STYLE dict, _style_axes helper,
_zscale_norm helper) for consistent visual style across all plots.

stage1.py, stage3.py, and stuck_shutters.py no longer import matplotlib
at module level — they call into plots.py only when plotting is needed.
Removes ~650 lines of duplicated plotting code from stage modules.
